requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
SQLAlchemy==2.0.35
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
aiohttp==3.10.10
httpx==0.27.0